+++ src/app/Router.js
@@ -3,10 +3,10 @@ import {Route, Switch, HashRouter} from 'react-router-dom';
 import App from './containers/App';
 import WelcomePage from './containers/WelcomePage';
 import SearchPage from './containers/SearchPage';
-import ErrorPage from './containers/ErrorPage';
+import RescanPage from './containers/RescanPage';
 import AppliancesPage from './containers/AppliancesPage';
 import DetectionLog from './containers/log/detectionLog';
-import HelpPage from "./containers/helpPage";
+import HelpPage from "./containers/HelpPage";
 
 const Router = () => (
     <HashRouter>
@@ -14,7 +14,7 @@ const Router = () => (
             <Route exact path="/" component={App}/>
             <Route exact path="/welcome" component={WelcomePage}/>
             <Route exact path="/search" component={SearchPage}/>
-            <Route exact path="/error" component={ErrorPage}/>
+            <Route exact path="/error" component={RescanPage}/>
             <Route exact path="/appliances" component={AppliancesPage}/>
             <Route exact path="/log" component={DetectionLog}/>
             <Route exact path="/help" component={HelpPage}/>
